First working version without working tests
