+++ tox.ini
@@ -2,8 +2,14 @@
 envlist = py26,py27,py33,py34,py35,pypy
 
 [testenv]
+deps =
+    nose
+
+changedir={toxinidir}/tests
 commands =
+    python test_coverage.py
     python test_profilehooks.py
+    python test_timeit_logger.py
 
 [testenv:py]
 commands =
